facenet/requirements.txt
opencv-python
scipy
pillow
gdown>=3.10.1
tf-keras